requirements.txt
discord.py
python-benedict